DAS plugins: Use a caret dependency on the analysis_server_plugin package
Also stop generating that version number. This code that used the
version of the analysis_server_plugin package in the Dart SDK was
defensive, but is now seen as overly restrictive; it does not allow
bug fixes to be used, to no benefit.

The only reason to bump the minimum version here is if there is a
change made to the API spec which is breaking, which requires a new
version of the package in order to communicate with DAS.

Change-Id: I13dab8135a8ec754aa993509656d871a547cb9cc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/446686
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>