[model] Remove unused remained code after a refactoring
The method `inferTypeFromConstraints` was merged with the shared
implementation of chooseTypes in one of the previous refactorings. The
method itself wasn't removed by an oversight. This CL addresses the
issue.

This CL also removes the stub implementation of `chooseTypes` from
`MiniAstOperations`, since the shared implementation of `chooseTypes` is
already provided and is in effect in all of the type systems, including
the one from the mini ast.

This is a follow-up for
https://dart-review.googlesource.com/c/sdk/+/446660

Change-Id: Ib15e1521444602cd79f093e221bbca81a459a191
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/446901
Commit-Queue: Chloe Stefantsova <cstefantsova@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>