[dart2js] Remove obsolete code from DCE
Remove an obsolete optimization: an instruction that has no users and
has the only effect of checking the the receiver is non-null, can be
removed when followed by another instruction with the same
null-checking effect.

With sound null safety, this rarely occurs. Most cases where this
would have fired are now explicit null checks that are not dead
because the second instruction is data-dependent of the check
instruction. There is similar code in instruction selection to omit
null checks, so we can rely on that.

The remaining cases are with back-to-back JS-fragments. This can be managed by adding a null check. Added a test case for an added null check.

Issue: #60327
Change-Id: Iace58bd944839166bc677d85bba2fe8d5c75da4c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/446483
Reviewed-by: Mayank Patke <fishythefish@google.com>
Commit-Queue: Stephen Adams <sra@google.com>