[cfe] Perform extension invocation inference on the original arguments
This changes how extension invocation is performed. Previously the inference was done on the lowered encoding, requiring the inference to bypass the synthesized extension type parameters and the extension receiver. The new approach uses the arguments directly for the source, avoiding a lot of complexity in the inference and aligning with the principle of only performing lowering _after_ inference.

Change-Id: I70973ee647b074589e1905c2993a846c15a70ce4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/447581
Commit-Queue: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>