Roll the latest dart_style into the SDK.
This only has one change, a bug fix:

https://github.com/dart-lang/dart_style/pull/1766

Prior to this change, if you have a file with an "@dart=" language
version comment, that version number won't be used to decide whether to
apply the 3.7 or 3.8 style changes. (It *does* correctly decide whether
to apply the 3.7 or older short style.) This fixes that.

It shouldn't have any noticeable effect in the SDK. The only mentions of
"@dart=3.7", "@dart=3.8", or "@dart=3.9" I see are in the formatter's
own tests.

Change-Id: I48e55637c7373e09bcf40befc174bebacecf8e9a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/447762
Reviewed-by: Ivan Inozemtsev <iinozemtsev@google.com>
Auto-Submit: Bob Nystrom <rnystrom@google.com>
Commit-Queue: Bob Nystrom <rnystrom@google.com>
Commit-Queue: Ivan Inozemtsev <iinozemtsev@google.com>