[front_end] Pass all diagnostic arguments to Message.arguments prior to conversion.
Previously, most entries in the `Message.arguments` map directly
reflected the values passed to `withArgumentsOld` (prior to converting
and the arguments into a form more suitable for display), but there
were some exceptions:

- For `name`, `name2`, `name3`, and `name4` parameters, the entry in
  `Message.arguments` was the value of the name after passing it
  through the function `demangleMixinApplicationName`.

- For `nameOKEmpty`, the entry in `Message.arguments` was the value of
  the name after substituting `(unnamed)` for the empty string.

- For `stringOKEmpty`, the entry in `Message.arguments` was the value
  of the name after substituting `(empty)` for the empty string.

This CL changes the generated `withArgumentsOld` functions so that all
the values in `Message.arguments` uniformly directly reflect the
values passed to `withArgumentsOld`.

In practice this has no user-visible effect, because the only use of
`Message.arguments` outside of testing is to convert parser
diagnostics to analyzer form, and none of the above conversions have
any effect on parser diagnostics. (This is because (a) mixin name
mangling happens after the parsing stage, and (b) the parser
diagnostics don't use `nameOKEmpty` or `stringOKEmpty` parameters.)

In follow-up CLs, I plan to rework the code generator for
`withArgumentsOld` so that it's easy to generate a parallel
`withArguments` function that takes named arguments. Starting from a
more uniform behavior will make that rework easier.

Change-Id: I6a6a69642e1d9b8277c4959d4a9c2d5ae0cd5b8f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/447627
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>