[analyzer] Hardcode literateApiEnabled to true
The constant `literateApiEnabled` was a temporary artifact of the
rollout of the analyzer's new literate API for diagnostic
reporting. Now that the literate API is enabled, it is no longer
needed.

Hardcoding this constant simplifies the code generation logic for
analyzer and linter diagnostics. This should help pave the way for
sharing more of the code generation logic with the CFE.

Change-Id: I6a6a6964c16db74613f6c7d9a3f8ee6789b38dda
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/447944
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>