[analyzer] trim trailing whitespace from diagnostic messages.
This is something the CFE already does as part of its diagnostic
message code generation; the analyzer should do it too.

Making this change to the analyzer now paves the way for unifying the
analyzer and CFE code for reading and interpreting `messages.yaml`
files.

Change-Id: I6a6a6964b04ef0a67f8077becb3726bafc40a5cb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/448228
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>