[analyzer] Use type names from messages.yaml when documenting diagnostic params.
Changes the code generator for analyzer diagnostic messages so that
when generating documentation for diagnostic message parameters, it
uses the shared type names (those used in `messages.yaml`) rather than
the analyzer-specific ones.

In practice the only difference right now is that `Type` will be used
instead of `DartType`.

This paves the way for future CLs which will unify some of the
analyzer and front_end logic for generating code from the
`messages.yaml` files. Once that code is unified, there will be
diagnostic messages with parameter types that the analyzer doesn't yet
support. This change ensures that it will still be able to generate
documentation for those messages.

Change-Id: I6a6a696471f530aa1f8ddb6f8f73beba6a0c70b7
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/448229
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>