[front_end] Make a script to add "parameters" entries to messages.yaml.
The new script, `pkg/front_end/tool/add_parameters_to_messages.dart`,
iterates through all the entries in `pkg/front_end/messages.yaml`. For
each parameter used in a `problemMessage` or `correctionMessage`, it
generates a `parameters` entry, with the appropriate parameter name
and type, and a documentation placeholder ("undocumented"). Future CLs
will fill in documentation strings and rename the parameters as needed
for clarity.

For messages that don't take any parameters, the script adds
`parameters: none` to the YAML source.

To make the script easier to write, a `yamlNode` field is added to the
`ErrorCodeInfo` class; this makes it easy for the script to find the
location in the YAML file where message is declared, so that it can
add the `parameters` entry.

Change-Id: I6a6a6964cbbd01bbe0cf7dbb1ef7944e0a61f6e9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/448250
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>