[front_end] Allow messages to have arbitrary parameter names.
Changes `ParsedPlaceholder.fromMatch` so that it no longer computes
the parameter's type by looking it up in a hardcoded table. Instead,
the parameter's type is instead obtained by looking it up by name in
the `ErrorCodeInfo.parameters` map.

This allows messages to have arbitrary parameter names. The hardcoded
table, `_templateParameterNameToType`, is removed since it is no
longer needed.

This also enabled some clean-up of the template compiler. I've split
it into three methods:

- `compile`, which compiles the entire template.

- `computeInterpolator`, which figures out the correct interpolation
  expression for a single parameter substitution.

- `interpolate`, which converts either the `problemMessage` or the
  `correctionMessage` into an interpolation string.

There is a small change to the generated code: in messages that use a
`TypeLabeler`, the `TypeLabeler` is declared and initialized just
before its first use, rather than at the top of the `withArguments`
method. This change makes the generation of the `withArguments` method
a little more straightforward, since it allows the body of
`withArguments` to be computed via a single scan through all the
placeholders. This change shouldn't have any user-visible effect.

Change-Id: I6a6a69647a9af5f32c3102b78d54431e290e0591
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/448237
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>