[front_end] Convert an error message to use named arguments.
Modifies the scanner logic that reports the `AsciiControlCharacter`
error message so that it invokes it using `withArguments` (which
accepts named arguments) rather than the old `withArgumentsOld` (which
accepts positional arguments). The name of the parameter accepted by
this error message has been changed to `character`, to reflect its
usage, and parameter documentation has been added.

This is an exercise to verify that the front end's new `withArguments`
error reporting infrastructure is fully functional. In follow-up CLs,
I plan to introduce some tools to partially automate the process of
updating front end code to use `withArguments` rather than
`withArgumentsOld`.

However, if anyone wishes to start migrating error reporting to the
new scheme before that, they are welcome to do so.

Change-Id: I6a6a6964fc87d6bf551ec4891c56e6106dd9c098
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/448239
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>