BREAKING CHANGE: Fix `parseIPv4Address` to not allow leading zeros.
Rewrite `parseIPv4Address` and `parseIPv6Address`
to not do so much unnecessary allocation and copying.

This is a breaking change if someone, somewhere,
depends on parsing IPv4 addresses with leading zeros.
Neither RFC 3986 or the WhatWG URL spec allows leading
zeros, so it wouldn't work anywhere else. Because of that,
it's not expected to be breaking in practice.

Fixes #61392.

BUG= https://dartbug.com/61392

Change-Id: I045353d5d1606795048e12ee12cd12b30dbf8ab3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/447260
Commit-Queue: Lasse Nielsen <lrn@google.com>
Reviewed-by: Ömer Ağacan <omersa@google.com>
Reviewed-by: Slava Egorov <vegorov@google.com>
Reviewed-by: Stephen Adams <sra@google.com>