analyzer: rename some variables in options_file_validator
I think it was me who renamed this variable
`sourceIsOptionsForContextRoot` earlier, but that was a complete
misunderstanding of it; the options file could be found anywhere,
maybe in the context root, maybe in a _different_ context root, maybe
in the parent of one, or a child directory of one. That is not what
this field indicates; it indicates whether the options file is being
analyzed as part of an include chain or not.

Additionally, `_TopLevelOptionValidator.pluginName` made no sense;
`sectionName` is a good name.

Change-Id: Ifc96e18d3a5e6b4988b1bf1ba4fb588605c73315
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/449077
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>