[analyzer] Ignore info can parse 'word' starting with underscore
Analyzer plugin warnings and lints can be ignored with
"// Ignore: plugin_name/lint_name".
Plugins are just packages and packages can start with an underscore,
so the ignore syntax should support that too, which it did not.

This CL changes that so it's supported.
It furthermore changes the extraction method (`ignoredElements`) from
sync* to be a regular sync method because it's so much easier to step
through. It's also faster though:

Below I've run this:

```dart
  Stopwatch stopwatch = Stopwatch()..start();
  CommentToken token = CommentToken(
    TokenType.STRING,
    // One of these:
    // "",
    // "// Ignore: foobar, baz",
    42,
  );
  for (int i = 0; i < 10000000; i++) {
    // ignoredElements1 or ignoredElements2 -- sync* vs sync.
    var list = token.ignoredElements2.toList();
    // One of these:
    // if (list.isNotEmpty) throw "Not empty.";
    // if (list.length != 2) throw "Not right size.";
  }
  print("Done in ${stopwatch.elapsed}");
```

as aot-compiled, run each 10 times and put the output through `ministat`:

Empty (''), sync* vs sync:

```
Difference at 95.0% confidence
        -0.629713 +/- 0.0755153
        -69.6007% +/- 8.34653%
        (Student's t, pooled s = 0.08037)
```

2 ignores ('// Ignore: foobar, baz'), sync* vs sync:

```
Difference at 95.0% confidence
        -1.17368 +/- 0.0544388
        -46.8828% +/- 2.17457%
        (Student's t, pooled s = 0.0579385)
```

Change-Id: I0c1abdc8001125e32bb5bddd579a187476f2fe31
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/449700
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>