[messages] Mark some messages as "pseudoShared".
A `pseudoShared` message is a front-end message that is generated by
code in the `_fe_analyzer_shared` package, and has a corresponding
analyzer message, but the error messages are not truly shared; the
analyzer message may have different text or parameters.

In the long term I plan to fully share all of these messages. However,
in the short term, this will allow the logic for front end error code
generation to be cleaned up, so that only messages that are truly
shared, or truly needed by `package:_fe_analyzer_shared` are emitted
to `pkg/_fe_analyzer_shared/lib/src/messages/codes_generated.dart`.

Change-Id: I6a6a6964bbb146afd41c39ab73b4bc8bbcc5b81c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/449607
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>