[analyzer/dead_code] Handle for-loop without updaters.
If the first dead node is the condition part of a for-loop,
`NullSafetyDeadCodeVerifier.flowEnd` normally adjusts the dead code
range to extend through the end of the loop's updaters list, to
improve the presentation of the warning. But if the loop has no
updaters, that would cause a crash, so to avoid the crash, the
adjustment is skipped.

Fixes https://github.com/dart-lang/sdk/issues/61428.

Change-Id: I6a6a6964a3866d5aa303ff921fa32abf7d4234bc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/449721
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Auto-Submit: Paul Berry <paulberry@google.com>