[messages] Only populate analyzerCodes for shared messages.
The `Code.analyzerCodes` field is only used for error codes that are
shared between the analyzer and the CFE. Error codes that are
generated into to `pkg/front_end` don't need them.

This paves the way for a follow-up CL in which I intend to change the
`analyzerCodes` field to an enum. It will avoid the need to put
unnecessary entries in the enum.

Change-Id: I6a6a69642d1ce012ebaf8fcc5ba11db8dc1b2535
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/449722
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>