[analyzer] Remove BaseDeprecatedMemberUseVerifier.reportError2.
During the element model migration, the method
`BaseDeprecatedMemberUseVerifier.reportError` had to be split into a
`reportError` and a `reportError2` version, to allow for a gradual
migration to the new element model. Now that the element model
migration is complete, there is no need to keep both methods.

See:

- https://dart-review.googlesource.com/c/sdk/+/396571, which added the
  `reportError2` method.

- https://dart-review.googlesource.com/c/sdk/+/393603, which migrated
  the linter's use of `BaseDeprecatedMemberUseVerifier`.

- https://dart-review.googlesource.com/c/sdk/+/402201, which migrated
  the analyzer's use of `BaseDeprecatedMemberUseVerifier`.

Change-Id: I6a6a69641640a2865922fc9b3142e8160e6d4ea6
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/450142
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>