[cfe] Add Resolver
This adds a Resolver class that encapsulates all use of the BodyBuilder. Furthermore, the external interface of BodyBuilder is extracted into the BodyBuilder interface and the BodyBuilderImpl implementation.

This is a step towards separating the BodyBuilder from the TypeInferrer. In future CLs the inferrer will be removed from BodyBuilder and Resolved will be responsible for first calling the BodyBuilder to create the uninferred AST nodes and then call the TypeInferrer to created the inferred and lowered AST nodes.


Change-Id: I198e2a65400555878d7b6c330fc6ef23be88851d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/449321
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>