[analyzer] Generalize DeprecatedMemberUseVerifier.
In a follow-up CL I plan to introduce a mechanism for detecting
problematic uses of elements marked `@experimental`. The logic for
doing this will be very similar to the logic currently used for
detecting problematic uses of elements marked `@deprecated`.

So to reduce the risk of bugs, this change generalizes the core
algorithm of `DeprecatedMemberUseVerifier` so that it will be able to
be used for both `@experimental` and `@deprecated`.

Generalization is done using two abstract strategy classes:

- `ElementUsageSet` abstracts the mechanism for detecting whether an
  element should not be used. In this this CL, there is one
  implementation to detect elements tagged with `@deprecated`:
  `DeprecatedElementUsageSet`.

- `ElementUsageReporter` abstracts the mechanism for reporting
  errors. In this CL, there are two implementations:
  `DeprecatedElementUsageReporter`, used by the analyzer for reporting
  "deprecated member use" warnings, and a private class
  `_DeprecatedElementUsageReporter`, used by the linter for reporting
  "deprecated member use from same package" lints.

The general algorithm is now in a pair of classes:

- `ElementUsageDetector`, for circumstances where it's necessary to
  detect _all_ usages of a certain set of elements.

- `ElementUsageFrontierDetector`, for circumstances where it's only
  necessary to detect usages of elements in the set from elements that
  are not in the set (i.e. usages that are on the frontier of the
  set). The name "frontier" was chosen by analogy to other uses of
  "frontier" in computer science and graph theory (e.g. "dominance
  frontier").

As a temporary measure, the class `BaseDeprecatedMemberUseVerifier`
remains, though it has been rewritten as a wrapper around
`ElementUsageDetector`. This is necessary because
`BaseDeprecatedMemberUseVerifier` is currently used in Google's
internal codebase. Once the internal use of
`BaseDeprecatedMemberUseVerifier` has been rewritten to use
`ElementUsageDetector`, `BaseDeprecatedMemberUseVerifier` will be
fully removed.

Note: in the future I intend to replace `ElementUsageDetector` and
`ElementUsageFrontierDetector` with visitors that use the linter's
registration mechanism, so that calls to their methods won't need to
be hardcoded into `BestPracticesVerifier`. But this will require some
infrastructure work, since the linter's registration mechanism is
currently only used for lints. So that will be in a future CL.

Change-Id: I6a6a69646f9eb49fb170d28ef2c926422f8091aa
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/450263
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>