[analysis_server] Update some more tests to handle EOL normalization
This sets `useLineEndingsForPlatform=true` for basic completion tests that now pass, but leave it set to `false` in `AbstractCompletionDriverTest` for now because there are a large number of failures in other tests that use that base class due to hard-coded offsets.

It also updates the format tests to support this (plus extracts some helpers to reduce the duplication).

Change-Id: Id2a0323b4605761d05e30bee89f7dccebb1f48c4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/451020
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>