[vm] Ignore experimental_member_use in tests of `@RecordUse`.
I will shortly be adding functionality to the analyzer to issue an
`experimental_member_use` warning when an element annotated with
`@experimental` is referenced. Since `@RecordUse` is currently
annotated with `@experimental`, we will need to ignore this warning
when testing it.

Tested: standard trybots
Change-Id: I6a6a6964293b7976c25121e4eb38bc9652501599
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/450973
Reviewed-by: Alexander Markov <alexmarkov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>