[dart2wasm] Move required js_util APIs out of js_util
Certain APIs that currently live in `dart:js_util` are used to implement
other behavior that we want to support, e.g. in `js_helper.dart` or
`js_interop_patch.dart`. We can move all of the logic we still need out
of `js_util` so that it can be updated without affecting the `js_util`
semantics. The goal is to replace the contents of `js_util` with
`UnsupportedError`s anyway.

Bug: #61550
CoreLibraryReviewExempt: code motion + doc update
Change-Id: I6a6a6964278650df57330fb088833bbdd9b43dd2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/451220
Reviewed-by: Srujan Gaddam <srujzs@google.com>